[ch32461] App blocked in country data changing
