+++ src_ts/components/common/mixins/uploads-mixin.ts
@@ -1,8 +1,16 @@
 import pmpEdpoints from '../../endpoints/endpoints.js';
 import {store, RootState} from '../../../redux/store';
-import {DECREASE_UNSAVED_UPLOADS, INCREASE_UPLOADS_IN_PROGRESS} from '../../../redux/actions/upload-status';
+import {
+  DECREASE_UNSAVED_UPLOADS,
+  INCREASE_UPLOADS_IN_PROGRESS,
+  RESET_UNSAVED_UPLOADS,
+  RESET_UPLOADS_IN_PROGRESS
+} from '../../../redux/actions/upload-status';
 import {Constructor} from '@unicef-polymer/etools-types';
 import {LitElement, property} from 'lit-element';
+import {openDialog} from '@unicef-polymer/etools-modules-common/dist/utils/dialog';
+import {translate} from 'lit-translate';
+
 /**
  * @polymer
  * @mixinFunction
@@ -26,6 +34,29 @@ function UploadsMixin<T extends Constructor<LitElement>>(baseClass: T) {
       if (state.uploadStatus!.uploadsInProgress !== this.uploadsInProgress) {
         this.uploadsInProgress = state.uploadStatus!.uploadsInProgress;
       }
+    }
+
+    public existsUploadsUnsavedOrInProgress(): boolean {
+      return Number(this.uploadsInProgress) > 0 || Number(this.unsavedUploads) > 0;
+    }
+
+    public async confirmLeaveUploadInProgress(): Promise<boolean> {
+      const confirmed = await openDialog({
+        dialog: 'are-you-sure',
+        dialogData: {
+          content: translate('LEAVE_UPLOAD_IN_PROGRESS'),
+          confirmBtnText: translate('LEAVE'),
+          cancelBtnText: translate('STAY')
+        }
+      }).then(({confirmed}) => {
+        return confirmed;
+      });
+
+      if (confirmed) {
+        store.dispatch({type: RESET_UNSAVED_UPLOADS});
+        store.dispatch({type: RESET_UPLOADS_IN_PROGRESS});
+      }
+      return confirmed;
     }
 
     public _onUploadStarted(e: any) {
